Update My custom GPT setup.md
updated with more unnecessary shit to get gpt-5 to stop complaining about policy. we can erase like all of my customizations if we get a logit bias and actual control over model set context memories. theyre terrible to work with due to how they've implemented writing to bio. it will straight up lie to you in the ui a lot of times because the systems are somewhat disconnected it seems... ui and memory managed memories... false advertising. one model will print a verbatim mem and the other shows a summarized or abridged mem. it's garbage. you can direct it, but who's got the time? it was built to become a mess from what i can tell.